refactor: extract server-auth
