+++ src/app/api/episodes/route.ts
@@ -1,28 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { createClient } from '@supabase/supabase-js';
 import { SupabaseAudioStorage } from '@/lib/supabase-storage';
-
-// Service role client for admin operations
-const supabaseServiceRole = createClient(
-  process.env.NEXT_PUBLIC_SUPABASE_URL!,
-  process.env.SUPABASE_SERVICE_ROLE_KEY!
-);
-
-// Helper to create user client
-const createUserClient = (token: string) => {
-  const supabaseUser = createClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-    {
-      global: {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      },
-    }
-  );
-  return supabaseUser;
-};
+import { createUserClient, supabaseServiceRole, getAuthFromRequest } from '@/lib/server-auth';
 
 export async function GET(request: NextRequest) {
   try {
@@ -125,19 +103,12 @@ export async function POST(request: NextRequest) {
     console.log('📨 Received POST request to /api/episodes');
     
     // Get auth token from header
-    const authHeader = request.headers.get('Authorization');
-    if (!authHeader?.startsWith('Bearer ')) {
-      console.log('❌ No valid auth header found');
+    const auth = await getAuthFromRequest(request);
+    if (!auth) {
+      console.log('❌ No valid auth found');
       return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
     }
-    
-    const token = authHeader.slice(7);
-    const { data: { user }, error: authError } = await supabaseServiceRole.auth.getUser(token);
-    
-    if (authError || !user) {
-      console.log('❌ Auth error:', authError);
-      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
-    }
+    const { user, token } = auth;
     
     console.log('✅ User authenticated:', user.id);
     console.log('🔍 User object:', { id: user.id, email: user.email, role: user.role });
@@ -202,17 +173,12 @@ export async function POST(request: NextRequest) {
 export async function DELETE(request: NextRequest) {
   try {
     // Get auth token from header
-    const authHeader = request.headers.get('Authorization');
-    if (!authHeader?.startsWith('Bearer ')) {
+    const auth = await getAuthFromRequest(request);
+    if (!auth) {
       return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
     }
     
-    const token = authHeader.slice(7);
-    const { data: { user }, error: authError } = await supabaseServiceRole.auth.getUser(token);
-    
-    if (authError || !user) {
-      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
-    }
+    const { user, token } = auth;
     
     const { searchParams } = new URL(request.url);
     const episodeId = searchParams.get('id');
