+++ src/app/api/topics/[id]/news/route.ts
@@ -1,24 +1,5 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { createClient } from '@supabase/supabase-js';
-
-// Service role client for privileged auth lookups
-const supabaseServiceRole = createClient(
-  process.env.NEXT_PUBLIC_SUPABASE_URL!,
-  process.env.SUPABASE_SERVICE_ROLE_KEY!
-);
-
-// Helper to create a RLS-respecting user client with a bearer token
-const createUserClient = (token: string) => {
-  return createClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-    {
-      global: {
-        headers: { Authorization: `Bearer ${token}` },
-      },
-    }
-  );
-};
+import { createUserClient, supabaseServiceRole, getAuthFromRequest } from '@/lib/server-auth';
 
 // Minimal RSS/Atom parser using fast-xml-parser
 async function fetchAndParseFeed(url: string) {
@@ -100,16 +81,12 @@ export async function GET(
   { params }: { params: { id: string } }
 ) {
   try {
-    const authHeader = request.headers.get('Authorization');
-    if (!authHeader?.startsWith('Bearer ')) {
+    const auth = await getAuthFromRequest(request);
+    if (!auth) {
       return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
     }
 
-    const token = authHeader.slice(7);
-    const { data: { user }, error: authError } = await supabaseServiceRole.auth.getUser(token);
-    if (authError || !user) {
-      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
-    }
+    const { user, token } = auth;
 
     const topicId = params.id;
     if (!topicId) {
